Removed localQuantity
This seems to have fixed a bug where the quantity was out of sync when changed in the cart.
To reproduce the bug, add some product to the cart. Then click on the product card within the cart, and add at least two more items, updating the total quantity in the cart. Now, if you increase or decrease the quantity from within the cart, the wrong quantity value would appear, and it seems to be because of the localQuantity. When the localQuantity state is removed, the bug disappears, and I can't see a new bug.

you would add a product to a cart, then go back to the product page, add at least 2 more of it.