Consistent Quantity in Product Page
When the user adds a product to their cart, and returns to the product page, the product's quantity is updated to match it accordingly in the product page. This also disables buttons when they shouldn't be used, and adds a saved state that checks if the user has made a change, and warns them if they attempt to leave prematurely.